# Optimize Address.Bytes() to inline fast path